Ability to set noise private key via environment
Moved some config code around to avoid circular
dependencies.
