+++ adapter/ph/Cargo.toml
@@ -11,7 +11,7 @@ blake3 = { version = "1.5.4" }
 bytes = { version = "1.7.1" }
 cbpf-rs = { path = "../../cbpf-rs" }
 chrono = "0.4.39"
-clap = { version = "4.5.7", features = ["derive"] }
+clap = { version = "4.5.7", features = ["derive", "env"] }
 crossbeam-epoch = { version = "0.9.18", optional = true }
 cslab = { path = "../../cslab" }
 curve25519-dalek = "4.1.3"
